Always configure intermediate parents when a child project is configured
When configure-on-demand is enabled, a project is evaluated only when required. In a nested
project structure, this can mean that a child project is configured without it's parent.
However, in a nested project structure, the parent project contributes a classloader that is used
by the child.

These 2 effects combined can result in:
- The parent project scope class loader is defined before the scope is locked.
  This violates the check enabled by `org.gradle.classloaderscope.strict`.
- In a composite build, plugins applied to the parent project may not be visible.

This commit fixes this by ensuring that intermediate parent projects are always configured
before a child project. (The root project is already configured before any children).

Fixes #14552